+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/background/database/AppDatabase.kt
@@ -7,7 +7,7 @@ import androidx.room.migration.Migration
 
 @Database(
     entities = [BackgroundJob::class, Notification::class],
-    version = 2
+    version = 1
 )
 abstract class AppDatabase : RoomDatabase() {
     abstract fun backgroundJobDao(): BackgroundJobDao