+++ www/OSHealthFitness.js
@@ -43,3 +43,7 @@ exports.setBackgroundJob = function (success, error, params) {
 exports.deleteBackgroundJob = function (success, error, params) {
     exec(success, error, 'OSHealthFitness', 'deleteBackgroundJob', [params]);
 };
+
+exports.listBackgroundJobs = function (success, error) {
+    exec(success, error, 'OSHealthFitness', 'listBackgroundJobs');
+};